Create maliciou_code_check
the code  searches for specific patterns in the HTML source code of the given URL to check for potential security vulnerabilities. It looks for script tags with a src attribute pointing to a known malicious site, form tags with an action attribute that submits data to a known malicious site, and input tags with a type of hidden and a name attribute of password, which could be used to steal a user's password.